Increase ActivityTimeout to 5min
Because cutscenes and vistas stop Mumble Link updates, having the
timeout too low would switch the setting to inactive during those, which
is undesired. Another solution would be to set the timeout really short,
so they play entirely with the inactive preset, but for vistas and
in-game cutscenes, that is not desired. The vistas could be helped by
detecting that the player co-ordinates are close to a vista, but that
does not help with the in-game cutscenes.